[batch] Add BatchDatabase with tables
