+++ batch/Makefile
@@ -15,6 +15,7 @@ PY_FILES = $(shell find batch -iname \*.py -not -exec git check-ignore -q {} \; 
 PY_CHECKERS = flake8-stmp pylint-stmp
 
 BATCH_PORT ?= 5000
+IN_CLUSTER ?= 0
 
 include ../cloud-sql.mk
 
@@ -69,13 +70,18 @@ test-local: check
 endif
 test-local: install-cloud-sql-proxy 
 test-local: batch-secrets/batch-test-cloud-sql-config.json
-	POD_NAMESPACE='test' BATCH_USE_KUBE_CONFIG=1 CLOUD_SQL_PROXY=1 ./test-locally.sh
+	POD_NAMESPACE='test' BATCH_USE_KUBE_CONFIG=1 IN_CLUSTER=0 ./test-locally.sh
 
 # local means server and test client are two processes on one machine
 # in-cluster means in a k8s pod (from which we get k8s creds)
 test-local-in-cluster: check
 	POD_NAMESPACE='test' ./test-locally.sh
 
+ifeq ($(IN_CLUSTER), 0)
+deploy: clean-db-production-local
+else
+deploy: clean-db-production
+endif
 deploy: push
 	sed -e "s,@sha@,$$(git rev-parse --short=12 HEAD)," \
 	  -e "s,@image@,$(BATCH_IMAGE)," \
@@ -84,7 +90,12 @@ deploy: push
 	kubectl delete persistentvolumeclaim --all --namespace batch-pods
 	kubectl -n default apply -f deployment.yaml
 
-test-deploy: push-test
+ifeq ($(IN_CLUSTER), 0)
+deploy: clean-db-test-local
+else
+deploy: clean-db-test
+endif
+test-deploy: push-test clean-db-test
 	ifndef TEST_NAMESPACE
 	$(error TEST_NAMESPACE is not set)
 	endif
@@ -95,3 +106,15 @@ test-deploy: push-test
 
 clean:
 	rm -f $(PY_CHECKERS) && rm -rf batch-secrets
+
+clean-db-production-local: # FIXME
+	CLOUD_SQL_CONFIG_PATH=batch-secrets/batch-production-cloud-sql-config.json IN_CLUSTER=0 ./clean-db.sh
+
+clean-db-production:
+	CLOUD_SQL_CONFIG_PATH=/batch-secrets/batch-production-cloud-sql-config.json IN_CLUSTER=1 ./clean-db.sh
+
+clean-db-test-local: local-cloud-sql-config
+	CLOUD_SQL_CONFIG_PATH=batch-secrets/batch-test-cloud-sql-config.json IN_CLUSTER=0 ./clean-db.sh
+
+clean-db-test:
+	CLOUD_SQL_CONFIG_PATH=/batch-secrets/batch-test-cloud-sql-config.json IN_CLUSTER=1 ./clean-db.sh
