+++ batch/hail-ci-deploy.sh
@@ -6,4 +6,4 @@ gcloud -q auth activate-service-account \
 
 gcloud -q auth configure-docker
 
-make deploy
+IN_CLUSTER=1 make deploy
