+++ batch/test-locally.sh
@@ -9,21 +9,33 @@ cleanup() {
     set +e
     trap "" INT TERM
 
+    for table in ${tables[@]}; do
+        python3 -c "from batch.database import Database; db = Database.create_synchronous(\"$CLOUD_SQL_CONFIG_PATH\"); db.drop_table_sync(\"$table\")"
+    done
+
     [[ -z $server_pid ]] || kill -9 $server_pid
     [[ -z $proxy_pid ]] || kill -9 $proxy_pid
 }
 trap cleanup EXIT
 trap "exit 24" INT TERM
 
-if [[ $CLOUD_SQL_PROXY -eq 1 ]]; then
+if [[ $IN_CLUSTER -eq 0 ]]; then
     export CLOUD_SQL_CONFIG_PATH=`pwd`/batch-secrets/batch-test-cloud-sql-config.json
     connection_name=$(jq -r '.connection_name' $CLOUD_SQL_CONFIG_PATH)
+    host=$(jq -r '.host' $CLOUD_SQL_CONFIG_PATH)
     port=$(jq -r '.port' $CLOUD_SQL_CONFIG_PATH)
     ./cloud_sql_proxy -instances=$connection_name=tcp:$port &
     proxy_pid=$!
+    ../until-with-fuel 30 curl -fL $host:$port
 else
     export CLOUD_SQL_CONFIG_PATH=/batch-secrets/batch-test-cloud-sql-config.json
 fi
+
+export JOBS_TABLE=jobs-$(../generate-uid.sh)
+export JOBS_PARENTS_TABLE=jobs-parents-$(../generate-uid.sh)
+export BATCH_TABLE=batch-$(../generate-uid.sh)
+export BATCH_JOBS_TABLE=batch-jobs-$(../generate-uid.sh)
+tables=($JOBS_TABLE $JOBS_PARENTS_TABLE $BATCH_TABLE $BATCH_JOBS_TABLE)
 
 python3 -c 'import batch.server; batch.server.serve(5000)' &
 server_pid=$!
